Improve names, types etc. js_only
