+++ assets/src/domain/wpPluginsPage/styles.scss
@@ -2,8 +2,8 @@
 	&__body {
 		min-width: 50%;
 		min-height: 80%;
-	}
-	iframe {
-		min-height: 40rem;
+		iframe {
+			min-height: 40rem;
+		}
 	}
 }
